--- app/_components/commodities/types.ts
@@ -1,15 +1,0 @@
-export type CommodityForm = {
-  commodityId: {
-    value: string;
-  };
-  system?: string;
-  includeOrbital?: boolean;
-  includePlanetary?: boolean;
-  includeOdyssey?: boolean;
-  includeFleetCarriers?: boolean;
-  maxLandingPadSize: string;
-  minSupply: number;
-  maxSupply: number;
-  minDemand: number;
-  maxDemand: number;
-};
