+++ app/_components/inputs/tests/Allegiances.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { useForm } from 'react-hook-form';
 import allegiances from '@/app/_lib/allegiance-list';
 import AllegiancesField from '../Allegiances';
-import { SystemForm } from '../../systems/types';
+import { SystemForm } from '@/app/_types/forms';
 
 describe('Allegiances Field', () => {
   beforeEach(() => {
