+++ app/_components/inputs/tests/Economies.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { useForm } from 'react-hook-form';
-import { SystemForm } from '../../systems/types';
+import { SystemForm } from '@/app/_types/forms';
 import EconomiesField from '../Economies';
 import economies from '@/app/_lib/economy-list';
