--- app/_components/systems/types.ts
@@ -1,14 +1,0 @@
-export type SystemForm = {
-  systemId?: string;
-  onlyPopulated?: boolean;
-  allegiance?: string;
-  government?: string;
-  economy?: string;
-  minorFaction?: string;
-  stationFilter?: string;
-  power?: string;
-  powerEffect?: string;
-  security?: string;
-  factionState?: string;
-  requiresPermit?: boolean;
-};
